// src/core/types.ts
export interface Data {
  league_logo: string;
  league_id: number;
  match_id: number;
  team_home_badge: string;
  match_hometeam_name: string;
  match_time: string;
  team_away_badge: string;
  match_awayteam_name: string;
  country_name: string;
  league_name: string;
  match_status: string;
  match_hometeam_score: number;
  match_awayteam_score: number;
  overall_league_position: number;
  team_name: string;
  overall_league_payed: number;
  overall_league_GF: number;
  overall_league_PTS: number;
  team_badge: string;
  overall_league_L: number;
  overall_league_D: number;
  overall_league_W: number;
  overall_league_GA: number;
  country_id: number;
  country_logo: string;
}